Added 3 new handlers, improved some old code.
- add handlers for add cooldown, foreign effect, itemgaininchat
- you can now see other players using buffs
- fixed buf of passive damage buffs acting as attack (berserk)
- fixed animations not being loaded if first frame is not 0
- improved pets, skillbook
- added EnumMap util to replace some arrays

+++ IO/Components/ScrollingNotice.cpp
@@ -54,7 +54,7 @@ namespace IO
 		if (active)
 		{
 			xpos -= 0.5f;
-			if (xpos < -notice.width())
+			if (xpos < static_cast<float>(-notice.width()))
 			{
 				xpos.set(800.0f);
 			}
